Job_application_automation/job_agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0
playwright>=1.44.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
reportlab>=4.0.0
python-docx>=1.1.0